build a general purpose converter + cxt->slf now available
the idea is that some formats are so close, they can be designed each
with a reader that yield data segment, and a writer that take these data
segments and yield lines to write in output file.
Combine a reader of a format with the writer of another, and you get a
fully working converter.